Cleaned up files before merging
Removed distortions folder as it was not from Livio or Adriaan
Fixed worst codacy errors
