+++ pycqed/analysis/analysis_toolbox.py
@@ -1572,16 +1572,6 @@ def find_min(x, y, return_fit=False, perc=30):
         return x_min, y_min
 
 
-def get_color_order(i, max_num, cmap='viridis'):
-    # take a blue to red scale from 0 to max_num
-    # uses HSV system, H_red = 0, H_green = 1/3 H_blue=2/3
-    # return colors.hsv_to_rgb(2.*float(i)/(float(max_num)*3.), 1., 1.)
-    print('It is recommended to use the updated function "get_color_cycle".')
-    if isinstance(cmap, str):
-        cmap = cm.get_cmap(cmap)
-    return cmap((i/max_num) % 1)
-
-
 def get_color_list(max_num, cmap='viridis'):
     '''Return an array of max_num colors take in even spacing from the
     color map cmap.'''
